backend/requirements.txt
# Core web framework dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
httpx==0.25.2

# Content extraction dependencies
beautifulsoup4==4.12.2
trafilatura==2.0.0
lxml>=5.3.0

# LLM API dependencies
google-generativeai==0.8.5

# Form parsing (to fix starlette warning)
python-multipart>=0.0.20

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Code quality tools
black==23.11.0
flake8==6.1.0
mypy==1.7.1
